requirements.txt
elasticsearch==2.4.0
robotframework==3.0.3